refactor main page and blog page
